% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{get_bounds}
\alias{get_bounds}
\title{Get Bounds for Outlier Detection}
\usage{
get_bounds(x = NULL, k = NULL)
}
\arguments{
\item{x}{A numeric vector. The data for which bounds are to be calculated.}

\item{k}{An optional numeric value. Controls the sensitivity of the outlier detection method.}
}
\value{
A numeric vector of bounds for the input data.
}
\description{
This function calculates the bounds for outlier detection using the LocScaleB method.
}
\examples{
get_bounds(c(1, 2, 3, 4, 5), k = 3)

}
\keyword{validation}
