% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{get_pricekg_bounds}
\alias{get_pricekg_bounds}
\title{Get Catch Price Bounds}
\usage{
get_pricekg_bounds(data = NULL, k = NULL)
}
\arguments{
\item{data}{A data frame containing columns: catch_taxon, catch_price_type, and price_kg_USD.}

\item{k}{A numeric value used in the LocScaleB function for outlier detection.}
}
\value{
A data frame with columns: catch_taxon, lower.low (lower bound), and upper.up (upper bound).
}
\description{
This function calculates the upper and lower bounds for catch price data based on catch taxon,
using the LocScaleB function for outlier detection.
}
\details{
The function filters out rows where catch_taxon is "0" or "no_catch". It then splits the data
based on catch_taxon and applies the get_bounds() function to each group's price_kg_USD values.
The resulting bounds are then combined and transformed back from log scale.
}
\keyword{validation}
