% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion.R
\name{ingest_surveys}
\alias{ingest_surveys}
\title{Ingest Surveys from Kobotoolbox into MongoDB}
\usage{
ingest_surveys(log_threshold = logger::DEBUG)
}
\arguments{
\item{log_threshold}{The (standard Apache log4j) log level used as a threshold for the logging infrastructure.
See \link[logger:log_levels]{logger::log_levels} for more details. Default is \code{logger::DEBUG}.}
}
\value{
This function performs side effects (downloads data, processes it, and uploads to MongoDB) and does not return a value.
}
\description{
This function downloads survey data from Kobotoolbox for a specific project, processes it into a tidy format,
and uploads the processed data to a MongoDB database. The function logs the progress at various stages.
Configuration parameters such as the URL, asset ID, username, and password are read from a configuration file
using \code{read_config()}.
}
\details{
The function performs the following steps:
\enumerate{
\item Sets the logging threshold to the specified level.
\item Reads configuration parameters using \code{read_config()}.
\item Logs the start of the data download process.
\item Downloads raw survey data from Kobotoolbox using \code{KoboconnectR::kobotools_kpi_data()}.
\item Checks for duplicate submission IDs to ensure data integrity.
\item Logs the start of data conversion to a tabular format.
\item Processes the raw data into a tidy tibble using \code{process_survey()}.
\item Logs the start of the data upload process to MongoDB.
\item Uploads the processed data to a MongoDB collection using \code{mdb_collection_push()}.
}
}
\examples{
\dontrun{
# Ensure that your configuration file is correctly set up with the necessary parameters.
ingest_surveys()
}
}
\keyword{workflow}
