% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peskas.malawi.data.pipeline-package.R
\docType{package}
\name{peskas.malawi.data.pipeline-package}
\alias{peskas.malawi.data.pipeline}
\alias{peskas.malawi.data.pipeline-package}
\title{peskas.malawi.data.pipeline: Functions to Implement the Malawi Small Scale Fisheries Data Pipeline}
\description{
This package implements the data and modelling pipelines underpinning the Peskas system for Malawi's small scale fisheries. It provides functions for data ingestion, preprocessing, validation, and storage, supporting efficient management and analysis of fisheries data.
}
\author{
\strong{Maintainer}: Lorenzo Longobardi \email{l.longobardi@cgiar.org} (\href{https://orcid.org/0000-0003-3126-7341}{ORCID})

Other contributors:
\itemize{
  \item WorldFish [copyright holder]
}

}
\keyword{internal}
