% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peskas.malawi.pipeline-package.R
\docType{package}
\name{peskas.malawi.pipeline-package}
\alias{peskas.malawi.pipeline}
\alias{peskas.malawi.pipeline-package}
\title{peskas.malawi.pipeline: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
