% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{validate_nboats}
\alias{validate_nboats}
\title{Validate Number of Boats}
\usage{
validate_nboats(data = NULL, k = NULL)
}
\arguments{
\item{data}{A data frame containing the \code{n_boats} column.}

\item{k}{A numeric value used in the LocScaleB function for outlier detection.}
}
\value{
A data frame with three columns:
\itemize{
\item \code{survey_id}: The unique identifier for each survey
\item \code{n_boats}: The original number of boats if valid, otherwise NA
\item \code{alert_n_boats}: A numeric value (3) indicating an outlier, or NA if the value is valid
}
}
\description{
This function validates the \code{n_boats} column in the provided dataset using the \code{alert_outlier}
function to detect outliers. If an outlier is detected, an alert is triggered and the \code{n_boats}
value is set to NA.
}
\keyword{validation}
