% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{validate_nfishers}
\alias{validate_nfishers}
\title{Validate Number of Fishers}
\usage{
validate_nfishers(data = NULL, k = NULL)
}
\arguments{
\item{data}{A data frame containing the \code{n_fishers} column.}

\item{k}{A numeric value used in the LocScaleB function for outlier detection.}
}
\value{
A data frame with three columns:
\itemize{
\item \code{survey_id}: The unique identifier for each survey
\item \code{n_fishers}: The original number of fishers if valid, otherwise NA
\item \code{alert_n_fishers}: A numeric value (2) indicating an outlier, or NA if the value is valid
}
}
\description{
This function validates the \code{n_fishers} column in the provided dataset using the \code{alert_outlier}
function to detect outliers. If an outlier is detected, an alert is triggered and the \code{n_fishers}
value is set to NA.
}
\keyword{validation}
